requirements.txt
discord.py
yt_dlp
python-dotenv